Add basic dashboard panel with disable override for rest of run button
