+++ schemas/disabled.json
@@ -1,0 +1,5 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema",
+  "type": "boolean",
+  "default": false
+}
