+++ src/dashboard/control/store.ts
@@ -1,0 +1,27 @@
+import { ReplicantModule, ReplicantTypes } from '@/browser_shared/replicant_store';
+import { Disabled } from '@/types/schemas';
+import Vue from 'vue';
+import Vuex, { Store } from 'vuex';
+import { getModule, Module, VuexModule } from 'vuex-module-decorators';
+
+Vue.use(Vuex);
+
+@Module({ name: 'OurModule' })
+class OurModule extends VuexModule {
+  // Helper getter to return all replicants.
+  get reps(): ReplicantTypes {
+    return this.context.rootState.ReplicantModule.reps;
+  }
+
+  get disabled(): Disabled {
+    return this.reps.disabled;
+  }
+}
+
+const store = new Store({
+  strict: process.env.NODE_ENV !== 'production',
+  state: {},
+  modules: { ReplicantModule, OurModule },
+});
+export default store;
+export const storeModule = getModule(OurModule, store);
