--- src/types/ExampleType.d.ts
@@ -1,3 +1,0 @@
-export interface ExampleType {
-  exampleProperty: string;
-}
