+++ src/types/schemas/disabled.d.ts
@@ -5,6 +5,4 @@
  * and run json-schema-to-typescript to regenerate this file.
  */
 
-export interface ExampleReplicant {
-	exampleProperty: string;
-}
+export type Disabled = boolean;
